+++ YarnSpinner/Types/EnumType.cs
@@ -33,9 +33,9 @@ namespace Yarn
     /// </summary>
     public class EnumType : TypeBase
     {
-        private readonly string name;
-        private readonly string description;
-        private readonly TypeBase rawType;
+        private string name;
+        private string description;
+        private TypeBase rawType;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="EnumType"/> class that
@@ -120,6 +120,10 @@ namespace Yarn
         {
             this._typeMembers.Add(name, member);
         }
+
+        internal void SetName(string name) => this.name = name;
+        internal void SetDescription(string description) => this.description = description;
+        internal void SetRawType(TypeBase rawType) => this.rawType = rawType;
 
         private static bool MethodEqualTo(Value a, Value b)
         {