+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - With this update, marker processors are able to register to handle this implicit marker, allowing games to customise the way that the name is handled.
 - Can now escape the implicit character separator `:`
   - technically can now escape any `:` but this only really has a user-facing impact on using it to escape the character name
+- Added `Yarn.Compiler.EnumTypeBuilder`, a helper for constructing new Yarn Spinner enum types in C#.
 
 ### Changed